refactor(platform): isolate GitHub data access
- Remove provider imports of task-facing platform modules.
- Share GitHub parsing and field queries in a lower-level module.
- Remove unused superseded metadata operations.
