+++ tests/unit/platform/pull-requests.test.ts
@@ -5,16 +5,8 @@ import os from 'node:os';
 import path from 'node:path';
 import { execFileSync, spawnSync } from 'node:child_process';
 
-import {
-  inspectPlatformPullRequestByNumber,
-  inspectGitHubIssueClosingChangeRequests,
-  normalizePullRequest,
-  selectExternalPullRequest,
-  selectPullRequest,
-  syncPlatformPullRequest,
-  syncPlatformPullRequestInLabels,
-  warningResultForPrimary
-} from '../../../lib/platform/pull-requests.ts';
+import { inspectPlatformPullRequestByNumber, selectExternalPullRequest, selectPullRequest, syncPlatformPullRequest, syncPlatformPullRequestInLabels, warningResultForPrimary } from '../../../lib/platform/pull-requests.ts';
+import { inspectGitHubIssueClosingChangeRequests, normalizePullRequest } from '../../../lib/platform/github-data.ts';
 import type { GitHubClient, RequestOptions } from '../../../lib/platform/github-client.ts';
 import { buildBoundFact, encodePrDeliveryFact } from '../../../lib/task/pr-delivery-fact.ts';
 import { resolveGitHubChangeRequestGitEvidence } from '../../../lib/platform/github-provider.ts';
